services/rag-service/pyproject.toml
[project]
name = "rag-service"
version = "0.1.0"
description = "RAG Service for DocIntel - Haystack-based document Q&A"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}

dependencies = [
    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.27.0",
    
    # RAG Framework
    "haystack-ai>=2.9.0",
    "qdrant-haystack>=6.0.0",
    "qdrant-client>=1.16.0",
    
    # LLM & Embeddings
    "litellm>=1.56.0",
    "ollama-haystack>=6.0.0",
    "sentence-transformers>=3.3.0",
    "transformers>=4.40.0",
    "accelerate>=0.26.0",
    "einops>=0.8.0",
    
    # Chunking
    "chonkie>=1.0.0",

    # Hybrid search (BM25 sparse vectors)
    "fastembed>=0.4.0",

    # Observability
    "langfuse>=2.57.0",
    "prometheus-fastapi-instrumentator>=7.0.0",
    
    # Templating
    "jinja2>=3.1.0",
    
    # Evaluation
    "datasets>=3.2.0",
    "ragas>=0.2.0",
    
    # Database
    "psycopg2-binary>=2.9.0",
    "sqlalchemy>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.1.0",
    "pytest-timeout>=2.2.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
